prefactor some functions out of the monolithic file
